split dbkz into its own file, euclid function top level, and minor syntactical changes
